removed some files and added some code safety
