+++ dev/com.ibm.ws.logging_fat/publish/servers/RealFlushServer/bootstrap.properties
@@ -1,2 +1,1 @@
 bootstrap.include=../testports.properties
-com.ibm.ws.logging.trace.specification=*=info:com.ibm.ws.collector.manager.*=debug:message.source.handler.*=debug
