--- dev/com.ibm.ws.logging_fat/test-applications/RealFlushTestApp/src/com/ibm/ws/logging/flush/fat/printFloatTests/PrintFloatTests.java
@@ -1,44 +1,0 @@
-package com.ibm.ws.logging.flush.fat.printFloatTests;
-
-import java.util.Set;
-import java.util.SortedMap;
-import java.util.logging.Logger;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.json.JsonObject;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.enterprise.inject.Produces;
-
-import com.ibm.websphere.filetransfer.FileServiceMXBean.MetaData;
-import com.ibm.websphere.logging.hpel.LogRecordContext;
-
-
-@ApplicationScoped
-@Path("/printFloatTests")
-public class PrintFloatTests {
-
-    private static final Logger MYlOGGER = Logger.getLogger(PrintFloatTests.class.getName());
-    
-    @GET
-    @Path("/printFloat")
-    public String printFloat() {
-    	
-    	float num = 222222222;
-    	System.out.print(num);
-    	
-        return "---- DONE ----";
-    }
-
-    @GET
-    @Path("/printlnFloat")
-    public String printlnFloat() {
-    	
-    	float num = 222222222;
-    	System.out.println(num);
-    	
-        return "---- DONE ----";
-    }
-}
