+++ dev/com.ibm.ws.logging_fat/test-applications/RealFlushTestApp/src/com/ibm/ws/logging/flush/fat/printTests/MyApplication.java
@@ -1,0 +1,33 @@
+/*******************************************************************************
+ * Copyright (c) 2019 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package com.ibm.ws.logging.flush.fat.printTests;
+
+import javax.ws.rs.core.Application;
+
+import com.ibm.json.java.JSONObject;
+
+import java.util.Set;
+import java.util.SortedMap;
+
+import javax.inject.Inject;
+import javax.enterprise.context.ApplicationScoped;
+import javax.json.JsonObject;
+import javax.ws.rs.ApplicationPath;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.Produces;;
+
+@ApplicationScoped
+@ApplicationPath("/")
+@Path("/")
+public class MyApplication extends Application {}
